Issue #2274: Move channel package out of navigation_overlay
