+++ app/src/main/java/org/mozilla/tv/firefox/channels/DefaultChannel.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.navigationoverlay.channels
+package org.mozilla.tv.firefox.channels
 
 import android.view.ViewGroup
 import android.widget.TextView
