+++ app/src/main/java/org/mozilla/tv/firefox/settings/SettingsFragment.kt
@@ -21,8 +21,8 @@ import org.mozilla.tv.firefox.R
 import org.mozilla.tv.firefox.architecture.FirefoxViewModelProviders
 import org.mozilla.tv.firefox.ext.forceExhaustive
 import org.mozilla.tv.firefox.ext.serviceLocator
-import org.mozilla.tv.firefox.navigationoverlay.channels.SettingsScreen
-import org.mozilla.tv.firefox.navigationoverlay.channels.SettingsTile
+import org.mozilla.tv.firefox.channels.SettingsScreen
+import org.mozilla.tv.firefox.channels.SettingsTile
 
 const val KEY_SETTINGS_TYPE = "KEY_SETTINGS_TYPE"
